Some MCH simulation fixes of mostly stupid bugs.
- Hit : Fix detElemId return type
- Hit : Entrance point was not recorded
- GeometryTest : Fix getNormalVector function

+++ Detectors/MUON/MCH/Simulation/include/MCHSimulation/Hit.h
@@ -31,7 +31,7 @@ class Hit : public ::o2::BasicXYZEHit<float>
   Point3D<float> entrancePoint() const { return GetPos(); }
   Point3D<float> exitPoint() const { return mExitPoint; }
 
-  float detElemId() const { return GetDetectorID(); }
+  short detElemId() const { return GetDetectorID(); }
 
  private:
   float mLength = {};
